Delete ToolWindowCommandTests suite to avoid add shell-specific logic to TestsInfrastructure project
